requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pyDes==2.0.1
PyExecJS==1.5.1
pymongo==3.11.3
requests==2.25.1
six==1.15.0
urllib3==1.26.4
